# docs/openApi.json
{
 "components": {
  "schemas": {
   "Error": {
    "description": "Standard JSON error",
    "properties": {
     "message": {
      "type": "string"
     },
     "statusCode": {
      "maximum": 600,
      "minimum": 100,
      "type": "integer"
     }
    },
    "required": [
     "message",
     "statusCode"
    ],
    "type": "object"
   },
...
   "viseh-api_types_FullUser": {
    "description": "User object",
    "nullable": true,
    "properties": {
     "email": {
      "description": "email address",
      "example": "alberthagenes@kemmer.name",
      "type": "string"
     },
     "id": {
      "description": "unique user id",
      "example": 1,
      "type": "integer"
     },
     "med_id": {
      "description": "unique medical id",
      "example": "12345abc",
      "type": "string"
     },
     "name": {
      "description": "first and last name",
      "example": "Loraine Stoltenberg",
      "type": "string"
     },
     "password": {
      "type": "string"
     },
     "phone_number": {
      "description": "phone number",
      "example": "995-399-6563",
      "type": "string"
     },
     "role": {
      "description": "role",
      "enum": [
       "user",
       "employee",
       "admin"
      ],
      "example": "user",
      "type": "string"
     }
    },
    "type": "object"
   },
...
   "viseh-api_types_User": {
    "description": "User object with values allowed to be patched",
    "nullable": true,
    "properties": {
     "email": {
      "type": "string"
     },
     "id": {
      "type": "integer"
     },
     "med_id": {
      "type": "string"
     },
     "name": {
      "type": "string"
     },
     "phone_number": {
      "type": "string"
     },
     "role": {
      "type": "string"
     }
    },
    "required": [
     ""
    ],
    "type": "object"
   },
   "viseh-api_types_UserLogin": {
    "properties": {
     "email": {
      "type": "string"
     },
     "password": {
      "type": "string"
     }
    },
    "type": "object"
   }
  }
 },
 "info": {
  "title": "messages",
  "version": "0.0.0"
 },
 "openapi": "3.0.0",
 "paths": {
...
  "/login": {
   "post": {
    "description": "Login.",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/viseh-api_types_UserLogin"
       }
      }
     }
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "default": {
      "description": ""
     }
    }
   }
  },
...
  "/register": {
   "post": {
    "description": "Register.",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/viseh-api_types_FullUser"
       }
      }
     }
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "default": {
      "description": ""
     }
    }
   }
  },
  "/users": {
   "get": {
    "description": "Get list of users.",
    "parameters": [
     {
      "description": "amount of items",
      "in": "query",
      "name": "items",
      "schema": {
       "type": "integer"
      }
     },
     {
      "description": "current page index",
      "in": "query",
      "name": "page",
      "schema": {
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/viseh-api_types_User"
         },
         "nullable": true,
         "type": "array"
        }
       }
      },
      "description": ""
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "default": {
      "description": ""
     }
    }
   }
  },
  "/users/search": {
   "get": {
    "description": "Search for user.",
    "parameters": [
     {
      "description": "user email",
      "in": "query",
      "name": "email",
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/viseh-api_types_User"
         },
         "nullable": true,
         "type": "array"
        }
       }
      },
      "description": ""
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "default": {
      "description": ""
     }
    }
   }
  },
  "/users/{uid}": {
   "get": {
    "description": "Get user by uid.",
    "parameters": [
     {
      "description": "id of the user",
      "in": "path",
      "name": "uid",
      "required": true,
      "schema": {
       "pattern": "\\d+",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/viseh-api_types_User"
        }
       }
      },
      "description": ""
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "default": {
      "description": ""
     }
    }
   },
   "patch": {
    "description": "Patch user.",
    "parameters": [
     {
      "description": "id of the user",
      "in": "path",
      "name": "uid",
      "required": true,
      "schema": {
       "pattern": "\\d+",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/viseh-api_types_User"
       }
      }
     }
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "default": {
      "description": ""
     }
    }
   },
   "put": {
    "description": "Update user.",
    "parameters": [
     {
      "description": "id of the user",
      "in": "path",
      "name": "uid",
      "required": true,
      "schema": {
       "pattern": "\\d+",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/viseh-api_types_User"
       }
      }
     }
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "401": {
      "content": {
       "application/json": {
        "schema": {
         "type": "string"
        }
       }
      },
      "description": ""
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      },
      "description": ""
     },
     "default": {
      "description": ""
     }
    }
   }
  }
 }
}